Migrate comment audit classes. Register audit provider, domain kinds, and import strategy from SpecimenModule.
