+++ api/src/org/labkey/api/util/PageFlowUtil.java
@@ -1419,7 +1419,7 @@ public class PageFlowUtil
         return '"';
     }
 
-    @Deprecated    // Use LinkBuilder directly - see PageFlowUtil.link(). 35 usages.
+    @Deprecated    // Use LinkBuilder directly - see PageFlowUtil.link(). 33 usages.
     public static String textLink(String text, URLHelper url)
     {
         return link(text).href(url).toString();
